+++ sistema/src/main/java/com/mateus/sistema/domain/Pais.java
@@ -21,7 +21,7 @@ public class Pais implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 	private String codigo;
 	private String sigla;
@@ -33,18 +33,18 @@ public class Pais implements Serializable {
 	public Pais() {
 	}
 
-	public Pais(Integer id, String nome, String codIBGE, String sigla) {
+	public Pais(Long id, String nome, String codIBGE, String sigla) {
 		this.id = id;
 		this.nome = nome;
 		this.codigo = codIBGE;
 		this.sigla = sigla;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
