+++ sistema/src/main/java/com/mateus/sistema/domain/PessoaEndereco.java
@@ -20,10 +20,10 @@ public class PessoaEndereco implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@Column(name = "pessoa_id")
-	private Integer pessoa;
+	private Long pessoa;
 
 	@ManyToOne
 	@JoinColumn(name = "endereco_id")
@@ -34,18 +34,18 @@ public class PessoaEndereco implements Serializable {
 	public PessoaEndereco() {
 	}
 
-	public PessoaEndereco(Integer id, Pessoa pessoa, Endereco endereco) {
+	public PessoaEndereco(Long id, Pessoa pessoa, Endereco endereco) {
 		this.id = id;
 		this.pessoa = pessoa.getId();
 		this.tipo = pessoa.getTipo().getCod();
 		this.endereco = endereco;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
